# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Generative Walking Route Planner API
  description: |
    A vibe-based walking route planner API for the IBM Dev Day "AI Demystified" Hackathon.
    
    This API generates walking routes optimized for user-defined "vibes" like greenery,
    safety, quietness, and walkability. It uses OpenStreetMap data with a custom
    Generative Cost Function: Cg = C_base × (1 - Quality_Score)
    
    ## Data Source
    OpenStreetMap (OSM) under ODbL license.
    
    ## Supported Vibes
    - **Greenery**: Parks, gardens, trees
    - **Blue Space**: Rivers, lakes, coastline
    - **Introvert Mode**: Quiet, peaceful areas
    - **Extrovert Mode**: Lively, bustling areas
    - **Safety Check**: Well-lit, safe streets
    - **Walkability**: Pedestrian-friendly paths
    
    ## Integration with watsonx Orchestrate
    Import this OpenAPI spec as a Custom Skill to enable natural language routing.
  version: 1.0.0
  contact:
    name: IBM Hackathon Team
servers:
  - url: https://638be3913b41.ngrok-free.app
    description: Production server via ngrok

paths:
  /route:
    post:
      operationId: generateWalkingRoute
      summary: Generate a walking route based on vibes
      description: |
        Generate a walking route optimized for the user's vibe preferences.
        
        **For watsonx Orchestrate Agent:**
        Map user intents to vibe parameters:
        - "nature walk" → greenery=0.9, blue_space=0.3
        - "safe evening walk" → safety_check=0.95
        - "peaceful stroll" → introvert_mode=0.8
        - "lively city walk" → extrovert_mode=0.9
        - "quick walk to X" → walkability=0.9
      tags:
        - Routing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
            examples:
              natureLoop:
                summary: Nature Loop Walk
                value:
                  mode: circular_loop
                  origin:
                    lat: 14.5547
                    lon: 121.0244
                  duration_minutes: 45
                  vibes:
                    greenery: 0.9
                    safety_check: 0.8
                    introvert_mode: 0.6
                    walkability: 0.8
              safePointToPoint:
                summary: Safe Evening Route
                value:
                  mode: point_to_point
                  origin:
                    lat: 14.5547
                    lon: 121.0244
                  destination:
                    lat: 14.5580
                    lon: 121.0280
                  vibes:
                    safety_check: 0.95
                    walkability: 0.9
      responses:
        '200':
          description: Successfully generated route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error during route generation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      operationId: healthCheck
      summary: Check API health status
      tags:
        - System
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /set-location:
    post:
      operationId: setLocation
      summary: Store location in session for later retrieval
      description: |
        Called by frontend when user clicks on map.
        Stores location coordinates with a session ID.
        The session_id can later be used with /execute to retrieve the location.
      tags:
        - Session Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - lat
                - lon
              properties:
                session_id:
                  type: string
                  description: Unique session identifier
                lat:
                  type: number
                  description: Latitude
                lon:
                  type: number
                  description: Longitude
      responses:
        '200':
          description: Location stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  /cached-regions:
    get:
      operationId: listCachedRegions
      summary: List cached geographic regions
      tags:
        - System
      responses:
        '200':
          description: List of cached regions
          content:
            application/json:
              schema:
                type: object
                properties:
                  regions:
                    type: array
                    items:
                      type: string

  /execute:
    post:
      operationId: executeRoute
      summary: Execute route generation for Orchestrate
      description: |
        Pure execution endpoint for watsonx Orchestrate.
        
        This endpoint receives clean execution parameters from Orchestrate
        and performs the routing algorithm without any strategic decisions.
        
        All strategic analysis (vibe mapping, algorithm selection, etc.)
        is done by Orchestrate before calling this endpoint.
        
        Always generates circular loop walks using Dijkstra's algorithm.
      tags:
        - Orchestrate Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
            examples:
              highGreenery:
                summary: High greenery request
                value:
                  origin:
                    lat: 14.5547
                    lon: 121.0244
                  duration_minutes: 45
                  vibes:
                    greenery: 0.9
                    introvert_mode: 0.7
                    safety_check: 0.6
                    walkability: 0.8
              safeEvening:
                summary: Safe evening walk
                value:
                  origin:
                    lat: 14.5547
                    lon: 121.0244
                  duration_minutes: 30
                  vibes:
                    safety_check: 0.95
                    walkability: 0.9
                    greenery: 0.4
      responses:
        '200':
          description: Successfully generated route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
        '400':
          description: Route generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteErrorResponse'

components:
  schemas:
    RouteRequest:
      type: object
      required:
        - origin
      properties:
        mode:
          type: string
          enum: [point_to_point, circular_loop]
          default: point_to_point
          description: |
            Routing mode:
            - point_to_point: Route from origin to destination
            - circular_loop: Round-trip from origin back to origin
        origin:
          $ref: '#/components/schemas/Coordinate'
        destination:
          $ref: '#/components/schemas/Coordinate'
          description: Required for point_to_point mode
        duration_minutes:
          type: integer
          minimum: 5
          maximum: 180
          default: 30
          description: Target walk duration for circular_loop mode
        vibes:
          $ref: '#/components/schemas/VibeWeights'
        no_go_zones:
          type: array
          items:
            $ref: '#/components/schemas/NoGoZone'
          description: Polygon areas to avoid
        include_narrative:
          type: boolean
          default: true
          description: Generate AI explanation of route choice

    Coordinate:
      type: object
      required:
        - lat
        - lon
      properties:
        lat:
          type: number
          format: float
          minimum: -90
          maximum: 90
          description: Latitude in decimal degrees
        lon:
          type: number
          format: float
          minimum: -180
          maximum: 180
          description: Longitude in decimal degrees
      example:
        lat: 14.5547
        lon: 121.0244

    VibeWeights:
      type: object
      description: |
        Vibe preferences as floats from 0.0 to 1.0.
        Higher values indicate stronger preference for that vibe.
      properties:
        greenery:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.5
          description: Preference for parks, trees, green spaces (0=don't care, 1=maximum)
        blue_space:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0
          description: Preference for water bodies like rivers, lakes
        introvert_mode:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0
          description: Preference for quiet, peaceful areas (avoids busy streets)
        extrovert_mode:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0
          description: Preference for lively, bustling areas with shops/cafes
        safety_check:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.5
          description: Preference for well-lit, safe streets
        walkability:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.7
          description: Preference for pedestrian-friendly paths

    NoGoZone:
      type: object
      required:
        - vertices
      properties:
        vertices:
          type: array
          minItems: 3
          items:
            $ref: '#/components/schemas/Coordinate'
          description: Polygon vertices defining area to avoid

    RouteResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        geojson:
          $ref: '#/components/schemas/GeoJSONFeatureCollection'
        metadata:
          $ref: '#/components/schemas/RouteMetadata'

    GeoJSONFeatureCollection:
      type: object
      properties:
        type:
          type: string
          default: FeatureCollection
        features:
          type: array
          items:
            $ref: '#/components/schemas/GeoJSONFeature'

    GeoJSONFeature:
      type: object
      properties:
        type:
          type: string
          default: Feature
        geometry:
          $ref: '#/components/schemas/GeoJSONGeometry'
        properties:
          type: object
          additionalProperties: true

    GeoJSONGeometry:
      type: object
      properties:
        type:
          type: string
          default: LineString
        coordinates:
          type: array
          items:
            type: array
            items:
              type: number
          description: Array of [longitude, latitude] pairs

    RouteMetadata:
      type: object
      properties:
        distance_meters:
          type: number
          description: Total route distance in meters
        estimated_duration_minutes:
          type: number
          description: Estimated walking time in minutes
        vibe_score:
          type: number
          minimum: 0
          maximum: 1
          description: Overall route quality score based on vibes
        vibe_breakdown:
          type: object
          additionalProperties:
            type: number
          description: Individual vibe scores for transparency
        transparency_narrative:
          type: string
          nullable: true
          description: AI-generated explanation of route choice
        algorithm_used:
          type: string
          description: Routing algorithm used (dijkstra or astar)
        nodes_explored:
          type: integer
          description: Number of graph nodes explored

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          default: healthy
        version:
          type: string
        graph_loaded:
          type: boolean
          description: Whether OSM graph is loaded
        cached_regions:
          type: array
          items:
            type: string

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        error_code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          nullable: true
          additionalProperties: true

    ExecuteRequest:
      type: object
      properties:
        origin:
          $ref: '#/components/schemas/Coordinate'
          description: Starting point coordinates. If not provided, session_id must be provided.
        session_id:
          type: string
          description: Session ID to retrieve location from frontend. Use this when origin is not provided directly.
        duration_minutes:
          type: integer
          minimum: 5
          maximum: 180
          default: 30
          description: Target walk duration in minutes. Always generates circular loop.
        vibes:
          $ref: '#/components/schemas/VibeWeights'
        no_go_zones:
          type: array
          items:
            $ref: '#/components/schemas/NoGoZone'
          description: Polygon areas to avoid during routing

    ExecuteResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        geojson:
          $ref: '#/components/schemas/GeoJSONFeatureCollection'
        metadata:
          type: object
          properties:
            distance_meters:
              type: number
              description: Total route distance in meters
            estimated_duration_minutes:
              type: number
              description: Estimated walking time in minutes
            vibe_score:
              type: number
              minimum: 0
              maximum: 1
              description: Overall route quality based on vibes
            vibe_breakdown:
              type: object
              properties:
                greenery:
                  type: number
                blue_space:
                  type: number
                quietness:
                  type: number
                liveliness:
                  type: number
                safety:
                  type: number
                walkability:
                  type: number
                overall:
                  type: number
        execution_details:
          type: object
          properties:
            algorithm:
              type: string
              default: dijkstra
              description: Pathfinding algorithm used
            nodes_explored:
              type: integer
              description: Number of graph nodes explored
            graph_size:
              type: string
              description: Size of OSM graph (nodes/edges)
            disjoint_percentage:
              type: number
              description: Percentage of return path that is unique (0-1)
            execution_time_ms:
              type: number
              description: Time taken to generate route in milliseconds

    ExecuteErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
          description: Error message for Orchestrate to display
        error_code:
          type: string
          description: Machine-readable error code
        suggestions:
          type: array
          items:
            type: string
          description: Suggestions for the user to try different parameters